Don't artificially inflate L0 score
Summary:
This turns out to be pretty bad because if we prioritize L0->L1 then L1 can grow artificially large, which makes L0->L1 more and more expensive. For example:
256MB @ L0 + 256MB @ L1 --> 512MB @ L1
256MB @ L0 + 512MB @ L1 --> 768MB @ L1
256MB @ L0 + 768MB @ L1 --> 1GB @ L1

....

256MB @ L0 + 10GB @ L1 --> 10.2GB @ L1

At some point we need to start compacting L1->L2 to speed up L0->L1.

I'm going to push this to a separate branch on github because we need this for Mongo+Rocks to be even able to import a giant dataset.

Test Plan: We'll test in prod

Reviewers: dynamike, sdong

Subscribers: dhruba, leveldb

Differential Revision: https://reviews.facebook.net/D38541