nvm: use direct io in nvm WritableFile
Use direct io in NVMWritableFile. The idea is to adapt the buffer size
in RocksDB write cache to the size of a physical block. This avoids
double buffering in the kernel; thus preventing the kernel to impede
future optimizations in RocksDB FTL.

Enabling direct IO requires to align data at 4KB (hardcoded page size
today). We will make this dynamic in the future.